requirements.txt
Faker
kafka-python
matplotlib
numpy
pandas
pyspark
python_dateutil
reverse_geocoder
